+++ public/index.js
@@ -1,3 +1,0 @@
-const { start } = window.wxtilesGl;
-
-start();