+++ src/layers/DebugTilesLayer.ts
@@ -1,13 +1,12 @@
-import { CompositeLayer } from '@deck.gl/core';
+import { Position, RGBAColor } from '@deck.gl/core';
+import { TextLayer, PathLayer } from '@deck.gl/layers';
 import { TileLayer } from '@deck.gl/geo-layers';
 import { TileLayerProps } from '@deck.gl/geo-layers/tile-layer/tile-layer';
-import { TextLayer } from '@deck.gl/layers';
-import { PathLayer } from '@deck.gl/layers';
-import { Position } from 'deck.gl';
-import { RenderSubLayers } from './IRenderSubLayers';
+
+import { RenderSubLayersProps } from './IRenderSubLayers';
 
 export interface DebugTilesLayerData {
-	color: [number, number, number, number?];
+	color: RGBAColor;
 }
 
 export interface DebugTilesLayerProps extends TileLayerProps<DebugTilesLayerData> {
@@ -19,29 +18,34 @@ export class DebugTilesLayer extends TileLayer<DebugTilesLayerData, DebugTilesLa
 		super(props);
 	}
 
-	renderSubLayers(args: RenderSubLayers) {
-		const {
-			x,
-			y,
-			z,
-			bbox: { west, south, east, north },
-		} = args.tile;
-		const { data } = this.props;
+	// getTileData(tile: any) {
+	// 	/* this is to test async fetching the data */
+	// 	return new Promise((resolve) => {
+	// 		setTimeout(() => {
+	// 			resolve(true);
+	// 		}, ~~(Math.random() * 1000 + 1000));
+	// 	});
+	// }
+
+	renderSubLayers(subProps: RenderSubLayersProps): any {
+		const { tile, id } = subProps;
+		const { x, y, z, bbox } = tile;
+		const { west, south, east, north } = bbox;
+		const { color } = this.props.data;
 		const subLayers = [
 			new TextLayer({
-				id: args.id + '-c',
-				visible: args.visible,
+				id: id + '-c',
 				data: [{}],
 				getPosition: () => [west + (east - west) * 0.05, north + (south - north) * 0.05], // if not ON TILE - visual issues occure
 				getText: () => x + '-' + y + '-' + z,
-				getColor: () => data.color,
+				getColor: color,
 				billboard: false,
 				getSize: 10,
 				getTextAnchor: 'start',
 			}),
+
 			new PathLayer({
-				id: args.id + '-b',
-				visible: args.visible,
+				id: id + '-b',
 				data: [
 					[
 						[west, north], // two (left and bottom) lines are enough to compose a square mesh
@@ -52,31 +56,20 @@ export class DebugTilesLayer extends TileLayer<DebugTilesLayerData, DebugTilesLa
 					],
 				],
 				getPath: (d) => d as Position[],
-				getColor: data.color,
+				getColor: color,
 				widthMinPixels: 1,
 			}),
 		];
 
-		// return new WxTileD({
-		// 	id: args.id + '-dd',
-		// 	visible: args.visible,
-		// 	data: subLayers,
-		// });
 		return subLayers;
 	}
 }
+
 DebugTilesLayer.layerName = 'DebugTilesLayer';
 DebugTilesLayer.defaultProps = {
-	tileSize: 256,
-	pickable: false,
 	data: { color: [255, 0, 0, 255] },
+	tileSize: 256,
 	maxZoom: 24,
 	minZoom: 0,
+	pickable: false,
 };
-
-class WxTileD extends CompositeLayer<any> {
-	renderLayers() {
-		return this.props.data;
-	}
-}
-WxTileD.layerName = 'WxTileD';