+++ src/layers/WxTileFill.ts
@@ -1,10 +1,13 @@
 import { BitmapLayer, BitmapLayerProps } from '@deck.gl/layers';
 import { picking, project32, Viewport } from '@deck.gl/core';
+import { UpdateStateInfo } from '@deck.gl/core/lib/layer';
+
 import { Texture2D } from '@luma.gl/webgl';
+
 import { ColorStyleStrict, HEXtoRGBA, UIntToColor } from '../utils/wxtools';
+
 import vs from '../shaders/bitmap-layer-vertex.vs';
 import fs from '../shaders/bitmap-layer-fragment.fs';
-import { UpdateStateInfo } from '@deck.gl/core/lib/layer';
 
 export interface WxTileFillData {
 	clutTextureUniform: Texture2D;
@@ -24,7 +27,7 @@ export class WxTileFill extends BitmapLayer<WxTileFillData, WxTileFillProps> {
 		super.updateState(a);
 		const { style } = this.props.data;
 		const fill = style.fill !== 'none';
-		const isolineColorUI = HEXtoRGBA(style.isolineColor);
+		const isolineColor = UIntToColor(style.isolineColor[0] === '#' ? HEXtoRGBA(style.isolineColor) : 0);
 		const isoline =
 			{
 				none: 0,
@@ -38,7 +41,7 @@ export class WxTileFill extends BitmapLayer<WxTileFillData, WxTileFillProps> {
 			bitmapTexture: this.props.data.imageTextureUniform,
 			fill,
 			isoline,
-			isolineColor: UIntToColor(isolineColorUI),
+			isolineColor,
 			desaturate,
 			transparentColor: transparentColor!.map((x) => x! / 255),
 			tintColor: tintColor!.slice(0, 3).map((x) => x / 255),
@@ -53,20 +56,20 @@ export class WxTileFill extends BitmapLayer<WxTileFillData, WxTileFillProps> {
 	draw(opts: any) {
 		const { uniforms, moduleParameters } = opts;
 		const { model, coordinateConversion, disablePicking } = this.state;
-		const { context } = this;
+		// const { context } = this;
 
 		if ((moduleParameters.pickingActive && disablePicking) || !model) {
 			return;
 		}
-		const viewport = context.viewport as Viewport & { center: number[] };
 
-		const camPos = viewport.getCameraPosition();
-		const camCent = viewport.center;
-		const sub = Math.sqrt((camPos[0] - camCent[0]) ** 2 + (camPos[1] - camCent[1]) ** 2 + (camPos[2] - camCent[2]) ** 2);
+		// const viewport = context.viewport as Viewport & { center: number[] };
+		// const camPos = viewport.getCameraPosition();
+		// const camCent = viewport.center;
+		// const sub = Math.sqrt((camPos[0] - camCent[0]) ** 2 + (camPos[1] - camCent[1]) ** 2 + (camPos[2] - camCent[2]) ** 2);
 		model.setUniforms({
 			...uniforms,
-			// shift: 1 / 255,
-			shift: sub / 50000,
+			shift: 1 / 255,
+			// shift: sub / 50000,
 			coordinateConversion,
 		});
 		this.state.model.draw(opts);