+++ src/wxtilesdeckgl.ts
@@ -1,0 +1,8 @@
+import './wxtilesdeckgl.css';
+
+export { setupWxTilesLib, setWxTilesLogging, createWxTilesLayerProps, CreateProps, WxServerVarsStyleType } from './libtools/libTools';
+
+export { WxTilesLayerManager, createDeckGlLayer } from './libtools/createDeckGlLayer';
+export { LibSetupObject, WXLOG } from './utils/wxtools';
+export { WxTilesLayer, WxTilesLayerProps } from './layers/WxTilesLayer';
+export { DebugTilesLayer } from './layers/DebugTilesLayer';