+++ watch-public.js
@@ -1,35 +1,26 @@
 const esbuild = require('esbuild');
-// const sassPlugin = require('esbuild-plugin-sass');
 const express = require('express');
-const path = require('path');
-
-let watchResponse;
-const disableHotReload = process.env.DISABLE_HOT_RELOAD === 'true';
 
 esbuild
 	.build({
-		entryPoints: ['src/dev-deckgl-index.ts', 'src/dev-mapbox-index.ts'],
+		entryPoints: ['src_example/dev-deckgl-index.ts'],
 		bundle: true,
-		// plugins: [sassPlugin()],
-		loader: {
-			'.ttf': 'base64',
-			'.woff': 'base64',
-			'.fs': 'text',
-			'.vs': 'text',
-		},
-		// target: 'es2017',
+		plugins: [],
+		loader: { '.woff': 'base64', '.fs': 'text', '.vs': 'text' },
 		format: 'iife',
-		outdir: 'public/wxtiles-gl',
-		globalName: 'wxtilesGl',
+		// https://www.stetic.com/market-share/browser/
+		target: ['es2020', 'chrome80', 'safari13', 'edge89', 'firefox70'],
+
+		outfile: 'public/wxtiles/wxtiles.js',
 		sourcemap: true,
-		// minify: false,
+		minify: true,
+
 		watch: {
 			onRebuild(error, result) {
 				if (error) {
 					console.error('watch build failed:', error);
 				} else {
 					console.log('rebuilded', new Date());
-					!disableHotReload && watchResponse && watchResponse.write('data: refresh\n\n');
 				}
 			},
 		},
@@ -37,14 +28,6 @@ esbuild
 	.then((result) => {
 		const app = express();
 		app.use(express.static('public'));
-		const publicPath = path.join(__dirname, 'public');
-
-		app.get('/mapbox', function (req, res) {
-			res.sendFile(path.join(publicPath, 'mapbox-index.html'));
-		});
-		app.get('/deckgl', function (req, res) {
-			res.sendFile(path.join(publicPath, 'deckgl-index.html'));
-		});
 
 		const PORT = 3005;
 
@@ -56,9 +39,9 @@ esbuild
 			});
 		});
 
-		const url = `http://0.0.0.0:${PORT}`;
+		const url = `http://localhost:${PORT}`;
 		app.listen(PORT, () => {
-			console.log(`See examples: \n${url}/mapbox\n${url}/deckgl`)
+			console.log(`See example: ${url}`);
 		});
 	})
 	.catch((e) => console.error(e.message));